Split from_str part
